+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ Release 1.6.1
 
 * Added a changelog.
 * Added the possibility to use weights in ``ihistogram``.
+* Added the function ``average_and_var`` to compute the average and variance in a single pass.
 * Documentation regarding the ``ddof`` keyword in many statistical wrongly stated that the default value was 1. This has been corrected. 
 
 Release 1.6